chore: create file for each component in form field generator
