+++ packages/client/src/v2-events/components/forms/FormFieldGenerator/GeneratedInputField.tsx
@@ -9,20 +9,11 @@
  * Copyright (C) The OpenCRVS Authors located at https://github.com/opencrvs/opencrvs-core/blob/master/AUTHORS.
  */
 
-/* eslint-disable */
-import React, { useCallback, useEffect } from 'react'
-import styled, { keyframes } from 'styled-components'
-import { Field, FieldProps, Formik, FormikProps } from 'formik'
-import { cloneDeep, isEqual, set, noop } from 'lodash'
-import {
-  WrappedComponentProps as IntlShapeProps,
-  MessageDescriptor,
-  useIntl
-} from 'react-intl'
+import React, { useCallback } from 'react'
+import { MessageDescriptor, useIntl } from 'react-intl'
 import {
   EventState,
   FieldConfig,
-  FieldType,
   FieldValue,
   FileFieldValue,
   FileFieldWithOptionValue,
@@ -34,7 +25,6 @@ import {
   isCountryFieldType,
   isDateFieldType,
   isDividerFieldType,
-  isFieldEnabled,
   isFileFieldType,
   isFileFieldWithOptionType,
   isLocationFieldType,
@@ -48,17 +38,13 @@ import {
   isTextFieldType,
   isNumberFieldType,
   isEmailFieldType,
-  isFieldVisible,
   isDataFieldType,
   EventConfig,
   EventIndex,
-  MetaFields,
-  AddressType,
   getDeclarationFields
 } from '@opencrvs/commons/client'
 import { TextArea } from '@opencrvs/components/lib/TextArea'
 import { SignatureUploader } from '@client/components/form/SignatureField/SignatureUploader'
-import { TEXT } from '@client/forms'
 import { InputField } from '@client/components/form/InputField'
 
 import {
@@ -82,55 +68,23 @@ import {
   Data,
   getFieldFromDataEntry
 } from '@client/v2-events/features/events/registered-fields/Data'
-import { useEventFormData } from '@client/v2-events/features/events/useEventFormData'
-import { useUserAddress } from '@client/v2-events/hooks/useUserAddress'
-import { Errors, getValidationErrorsForForm } from './validation'
-import {
-  evalExpressionInFieldDefinition,
-  FIELD_SEPARATOR,
-  getDependentFields,
-  handleDefaultValue,
-  hasDefaultValueDependencyInfo,
-  makeDatesFormatted,
-  makeFormFieldIdFormikCompatible
-} from './utils'
-import { File } from './inputs/FileInput/FileInput'
-import { FileWithOption } from './inputs/FileInput/DocumentUploaderWithOption'
+import { File } from '../inputs/FileInput/FileInput'
+import { FileWithOption } from '../inputs/FileInput/DocumentUploaderWithOption'
+import { makeFormikFieldIdsOpenCRVSCompatible } from './utils'
 
-const fadeIn = keyframes`
-  from { opacity: 0; }
-  to { opacity: 1; }
-`
-
-const FormItem = styled.div<{
-  ignoreBottomMargin?: boolean
-}>`
-  animation: ${fadeIn} 500ms;
-  margin-bottom: ${({ ignoreBottomMargin }) =>
-    ignoreBottomMargin ? '0px' : '22px'};
-`
-
-/*
- * Formik has a feature that automatically nests all form keys that have a dot in them.
- * Because our form field ids can have dots in them, we temporarily transform those dots
- * to a different character before passing the data to Formik. This function unflattens
- */
-function makeFormFieldIdsFormikCompatible<T>(data: Record<string, T>) {
-  return Object.fromEntries(
-    Object.entries(data).map(([key, value]) => [
-      makeFormFieldIdFormikCompatible(key),
-      value
-    ])
-  )
-}
-
-function makeFormikFieldIdsOpenCRVSCompatible<T>(data: Record<string, T>) {
-  return Object.fromEntries(
-    Object.entries(data).map(([key, value]) => [
-      key.replaceAll(FIELD_SEPARATOR, '.'),
-      value
-    ])
-  )
+interface ExposedProps {
+  fields: FieldConfig[]
+  id: string
+  fieldsToShowValidationErrors?: FieldConfig[]
+  setAllFieldsDirty: boolean
+  onChange: (values: EventState) => void
+  formData: Record<string, FieldValue>
+  requiredErrorMessage?: MessageDescriptor
+  onUploadingStateChanged?: (isUploading: boolean) => void
+  initialValues?: EventState
+  eventConfig?: EventConfig
+  declaration?: EventState
+  readonlyMode?: boolean
 }
 
 interface GeneratedInputFieldProps<T extends FieldConfig> {
@@ -166,7 +120,7 @@ interface GeneratedInputFieldProps<T extends FieldConfig> {
   readonlyMode?: boolean
 }
 
-const GeneratedInputField = React.memo(
+export const GeneratedInputField = React.memo(
   <T extends FieldConfig>({
     fieldDefinition,
     onChange,
@@ -537,300 +491,3 @@ const GeneratedInputField = React.memo(
 )
 
 GeneratedInputField.displayName = 'MemoizedGeneratedInputField'
-
-type FormData = Record<string, FieldValue>
-
-function mapFieldsToValues(
-  fields: FieldConfig[],
-  formData: FormData,
-  meta: MetaFields
-) {
-  return fields.reduce((memo, field) => {
-    const fieldInitialValue = handleDefaultValue(field, formData, meta)
-    return { ...memo, [field.id]: fieldInitialValue }
-  }, {})
-}
-interface ExposedProps {
-  fields: FieldConfig[]
-  id: string
-  fieldsToShowValidationErrors?: FieldConfig[]
-  setAllFieldsDirty: boolean
-  onChange: (values: EventState) => void
-  formData: Record<string, FieldValue>
-  requiredErrorMessage?: MessageDescriptor
-  onUploadingStateChanged?: (isUploading: boolean) => void
-  initialValues?: EventState
-  eventConfig?: EventConfig
-  eventDeclarationData?: EventState
-  readonlyMode?: boolean
-}
-
-type AllProps = ExposedProps &
-  IntlShapeProps &
-  FormikProps<EventState> & {
-    className?: string
-  }
-
-class FormSectionComponent extends React.Component<AllProps> {
-  componentDidUpdate(prevProps: AllProps) {
-    const userChangedForm = !isEqual(this.props.values, prevProps.values)
-    const sectionChanged = prevProps.id !== this.props.id
-
-    if (userChangedForm) {
-      prevProps.onChange(this.props.values)
-    }
-
-    if (sectionChanged) {
-      prevProps.resetForm()
-      if (this.props.setAllFieldsDirty) {
-        this.showValidationErrors(this.props.fields)
-      } else if (
-        this.props.fieldsToShowValidationErrors &&
-        this.props.fieldsToShowValidationErrors.length > 0
-      ) {
-        this.showValidationErrors(this.props.fieldsToShowValidationErrors)
-      }
-    }
-  }
-
-  async componentDidMount() {
-    if (this.props.setAllFieldsDirty) {
-      this.showValidationErrors(this.props.fields)
-    }
-
-    if (window.location.hash) {
-      setTimeout(() => {
-        const newScroll = document.documentElement.scrollTop - 100
-        window.scrollTo(0, newScroll)
-
-        const focusedElementId = window.location.hash.replace('#', '')
-        let focusedElement = document.querySelector(
-          `input[id*="${focusedElementId}"]`
-        ) as HTMLElement
-        if (focusedElement === null) {
-          // Handling for Select
-          focusedElement = document.querySelector(
-            `${window.location.hash} input`
-          ) as HTMLElement
-          focusedElement && focusedElement.focus()
-        } else {
-          // Handling for Input
-          focusedElement && focusedElement.focus()
-        }
-      }, 0)
-    }
-  }
-
-  showValidationErrors(fields: FieldConfig[]) {
-    const touched = fields.reduce((memo, field) => {
-      return { ...memo, [field.id]: true }
-    }, {})
-
-    void this.props.setTouched(touched)
-  }
-
-  setFieldValuesWithDependency = (
-    fieldName: string,
-    value: FieldValue | undefined
-  ) => {
-    const updatedValues = cloneDeep(this.props.values)
-    set(updatedValues, fieldName, value)
-
-    if (fieldName === 'country') {
-      set(
-        updatedValues,
-        'addressType',
-        value === (window.config.COUNTRY || 'FAR')
-          ? AddressType.DOMESTIC
-          : AddressType.INTERNATIONAL
-      )
-    }
-    const updateDependentFields = (fieldName: string) => {
-      const dependentFields = getDependentFields(this.props.fields, fieldName)
-      for (const field of dependentFields) {
-        if (
-          !field.defaultValue ||
-          !hasDefaultValueDependencyInfo(field.defaultValue)
-        ) {
-          continue
-        }
-
-        updatedValues[field.id] = evalExpressionInFieldDefinition(
-          field.defaultValue.expression,
-          { $form: updatedValues }
-        )
-        updateDependentFields(field.id)
-      }
-    }
-    updateDependentFields(fieldName)
-
-    void this.props.setValues(updatedValues)
-  }
-
-  resetDependentSelectValues = (fieldName: string) => {
-    const fields = this.props.fields
-    const fieldsToReset = fields.filter(
-      (field) => field.type === TEXT && field.dependsOn?.includes(fieldName)
-    )
-
-    fieldsToReset.forEach((fieldToReset) => {
-      void this.props.setFieldValue(fieldToReset.id, '')
-      this.resetDependentSelectValues(fieldToReset.id)
-    })
-  }
-
-  render() {
-    const {
-      values,
-      fields: fieldsWithDotIds,
-      touched,
-      intl,
-      className,
-      eventDeclarationData,
-      readonlyMode
-    } = this.props
-
-    const language = this.props.intl.locale
-
-    const errors = makeFormFieldIdsFormikCompatible(
-      this.props.errors as unknown as Errors
-    )
-
-    const fields = fieldsWithDotIds.map((field) => ({
-      ...field,
-      id: makeFormFieldIdFormikCompatible(field.id)
-    }))
-    const valuesWithFormattedDate = makeDatesFormatted(fieldsWithDotIds, values)
-
-    return (
-      <section className={className}>
-        {fields.map((field) => {
-          let error: string
-          const fieldErrors = errors[field.id] && errors[field.id].errors
-
-          if (fieldErrors && fieldErrors.length > 0) {
-            const [firstError] = fieldErrors
-            error = intl.formatMessage(firstError.message)
-          }
-
-          const formData = makeFormikFieldIdsOpenCRVSCompatible(
-            valuesWithFormattedDate
-          )
-
-          const allData = { ...formData, ...eventDeclarationData }
-
-          if (!isFieldVisible(field, allData)) {
-            return null
-          }
-
-          const isDisabled = !isFieldEnabled(field, allData)
-
-          return (
-            <FormItem
-              key={`${field.id}${language}`}
-              ignoreBottomMargin={field.type === FieldType.PAGE_HEADER}
-            >
-              <Field name={field.id}>
-                {}
-                {(formikFieldProps: FieldProps<any>) => {
-                  return (
-                    <GeneratedInputField
-                      fieldDefinition={field}
-                      resetDependentSelectValues={
-                        this.resetDependentSelectValues
-                      }
-                      setFieldValue={this.setFieldValuesWithDependency}
-                      {...formikFieldProps.field}
-                      disabled={isDisabled}
-                      error={isDisabled ? '' : error}
-                      eventConfig={this.props.eventConfig}
-                      fields={fields}
-                      formData={allData}
-                      readonlyMode={readonlyMode}
-                      touched={touched[field.id] ?? false}
-                      values={values}
-                      onUploadingStateChanged={
-                        this.props.onUploadingStateChanged
-                      }
-                    />
-                  )
-                }}
-              </Field>
-            </FormItem>
-          )
-        })}
-      </section>
-    )
-  }
-}
-
-export const FormFieldGenerator: React.FC<ExposedProps> = React.memo(
-  (props) => {
-    const { eventConfig, formData, fields, eventDeclarationData } = props
-
-    const intl = useIntl()
-    const { setAllTouchedFields, touchedFields: initialTouchedFields } =
-      useEventFormData()
-    const nestedFormData = makeFormFieldIdsFormikCompatible(formData)
-
-    const onChange = (values: EventState) => {
-      props.onChange(makeFormikFieldIdsOpenCRVSCompatible(values))
-    }
-    const user = useUserAddress()
-
-    const initialValues = makeFormFieldIdsFormikCompatible<FieldValue>({
-      ...mapFieldsToValues(props.fields, nestedFormData, { $user: user }),
-      ...props.initialValues
-    })
-
-    return (
-      <Formik<EventState>
-        enableReinitialize={true}
-        initialTouched={initialTouchedFields}
-        initialValues={initialValues}
-        validate={(values) =>
-          getValidationErrorsForForm(
-            fields,
-            makeFormikFieldIdsOpenCRVSCompatible(values)
-          )
-        }
-        validateOnMount={true}
-        onSubmit={noop}
-      >
-        {(formikProps) => {
-          const { touched } = formikProps
-
-          useEffect(() => {
-            /**
-             * Because 'enableReinitialize' prop is set to 'true' above, whenver initialValue changes,
-             * formik lose track of touched fields. This is a workaround to save all the fields that
-             * have been touched for once during the form manipulation. So that we can show validation
-             * errors for all fields that have been touched.
-             */
-            if (
-              Object.keys(touched).length > 0 &&
-              !isEqual(touched, initialTouchedFields) &&
-              Object.keys(touched).some((key) => !(key in initialTouchedFields))
-            ) {
-              setAllTouchedFields({
-                ...initialTouchedFields,
-                ...touched
-              })
-            }
-          }, [touched])
-          return (
-            <FormSectionComponent
-              {...props}
-              {...formikProps}
-              eventConfig={eventConfig}
-              eventDeclarationData={eventDeclarationData}
-              formData={nestedFormData}
-              intl={intl}
-              onChange={onChange}
-            />
-          )
-        }}
-      </Formik>
-    )
-  }
-)
